+++ data-raw/adding-logo.R
@@ -1,0 +1,5 @@
+fs::file_copy("~/Pictures/logos/BLASTr/BLASTr-logo.svg", "data-raw/")
+
+fs::file_copy("~/Pictures/logos/BLASTr/BLASTr-logo.png", "data-raw/")
+
+usethis::use_logo("data-raw/BLASTr-logo.png")