+++ tests/testthat/test-get_tax_id.R
@@ -1,5 +1,7 @@
-test_that("Get single taxid", {
+testthat::test_that("Get single taxid", {
+  testthat::skip_on_cran()
   testthat::skip_if_offline()
+
   res <- get_tax_by_taxID(organisms_taxIDs = "63221", verbose = FALSE)
 
   testthat::expect_s3_class(res, "tbl_df")
@@ -11,7 +13,8 @@ test_that("Get single taxid", {
   testthat::expect_equal(ncol(res), 13L)
 })
 
-test_that("Get single taxid empty df", {
+testthat::test_that("Get single taxid empty df", {
+  testthat::skip_on_cran()
   testthat::skip_if_offline()
   res <- get_tax_by_taxID(organisms_taxIDs = "10000000", verbose = FALSE)
 
@@ -22,7 +25,8 @@ test_that("Get single taxid empty df", {
   testthat::expect_equal(ncol(res), 13L)
 })
 
-test_that("Get parallel Single taxid", {
+testthat::test_that("Get parallel Single taxid", {
+  testthat::skip_on_cran()
   testthat::skip_if_offline()
   res <- parallel_get_tax(organisms_taxIDs = "63221", verbose = FALSE)
 
@@ -31,7 +35,8 @@ test_that("Get parallel Single taxid", {
   testthat::expect_true(stringr::str_detect(res$Sci_name, "Homo sapiens"))
 })
 
-test_that("Get parallel future Multiple taxid", {
+testthat::test_that("Get parallel Multiple taxid", {
+  testthat::skip_on_cran()
   testthat::skip_if_offline()
   res <- parallel_get_tax(
     organisms_taxIDs = c("63221", "10000000"),
@@ -46,12 +51,29 @@ test_that("Get parallel future Multiple taxid", {
 })
 
 
-test_that("Get parallel single-thread Multiple taxid", {
+testthat::test_that("Get parallel single-thread Multiple taxid", {
+  testthat::skip_on_cran()
   testthat::skip_if_offline()
   res <- parallel_get_tax(
     organisms_taxIDs = c("63221", "10000000"),
     total_cores = 2,
     retry_times = 2,
+    verbose = FALSE
+  )
+
+  testthat::expect_equal(res$query_taxID, "63221")
+
+  testthat::expect_true(stringr::str_detect(res$Sci_name, "Homo sapiens"))
+})
+
+
+testthat::test_that("`parallel_get_tax()` with `retry_times = 0` ", {
+  testthat::skip_on_cran()
+  testthat::skip_if_offline()
+  res <- parallel_get_tax(
+    organisms_taxIDs = c("63221", "10000000"),
+    total_cores = 1,
+    retry_times = 0,
     verbose = FALSE
   )